requirements.txt
# external requirements
flask
flask-cors
cianparser
scikit-learn
pandas
boto3
python-dotenv
PyGithub
flask_httpauth
catboost==1.0.4
numpy==1.21.0
gunicorn